Enable mailchimp settings per organization

+++ src/Controller/Admin/OrganizationsController.php
@@ -18,7 +18,11 @@ class OrganizationsController extends AppController
      */
     public function index()
     {
-        $organizations = $this->paginate($this->Organizations);
+        $organizations = $this->Organizations->find('all')
+            ->matching('Users', function($q) {
+                return $q->where(['Users.id' => $this->loggedInUser['id']]);
+            });
+        $organizations = $this->paginate($organizations);
 
         $this->set(compact('organizations'));
         $this->set('_serialize', ['organizations']);
@@ -76,9 +80,17 @@ class OrganizationsController extends AppController
      */
     public function edit($id = null)
     {
-        $organization = $this->Organizations->get($id, [
-            'contain' => []
-        ]);
+        $organization = $this->Organizations->findById($id)
+            ->matching('Users', function($q) {
+                return $q->where(['Users.id' => $this->loggedInUser['id']]);
+            })
+            ->first();
+
+        if (!$organization) {
+            $this->Flash->error(__('The organization could not be found. Please, select an existing organization or create a new one.'));
+            return $this->redirect(['action' => 'index']);
+        }
+
         if ($this->request->is(['patch', 'post', 'put'])) {
             $organization = $this->Organizations->patchEntity($organization, $this->request->data);
             if ($this->Organizations->save($organization)) {
@@ -94,6 +106,44 @@ class OrganizationsController extends AppController
     }
 
     /**
+     * Configure mailchimp method
+     *
+     * @param string|null $id Organization id.
+     * @return \Cake\Network\Response|void Redirects on successful edit, renders view otherwise.
+     * @throws \Cake\Network\Exception\NotFoundException When record not found.
+     */
+    public function configureMailchimp($id = null)
+    {
+        if (!$id) {
+            $id = $this->currentOrganization['id'];
+        }
+
+        $organization = $this->Organizations->findById($id)
+            ->matching('Users', function($q) {
+                return $q->where(['Users.id' => $this->loggedInUser['id']]);
+            })
+            ->first();
+
+        if (!$organization) {
+            $this->Flash->error(__('The organization could not be found. Please, select an existing organization or create a new one.'));
+            return $this->redirect(['action' => 'index']);
+        }
+        
+        if ($this->request->is(['patch', 'post', 'put'])) {
+            $organization = $this->Organizations->patchEntity($organization, $this->request->data);
+            if ($this->Organizations->save($organization)) {
+                $this->Flash->success(__('The organization has been saved.'));
+                return $this->redirect(['action' => 'index']);
+            } else {
+                $this->Flash->error(__('The organization could not be saved. Please, try again.'));
+            }
+        }
+        $this->set(compact('organization'));
+        $this->set('_serialize', ['organization']);
+        $this->viewBuilder()->template('configure_mailchimp');
+    }
+
+    /**
      * Delete method
      *
      * @param string|null $id Organization id.
@@ -103,7 +153,17 @@ class OrganizationsController extends AppController
     public function delete($id = null)
     {
         $this->request->allowMethod(['post', 'delete']);
-        $organization = $this->Organizations->get($id);
+        $organization = $this->Organizations->findById($id)
+            ->matching('Users', function($q) {
+                return $q->where(['Users.id' => $this->loggedInUser['id']]);
+            })
+            ->first();
+
+        if (!$organization) {
+            $this->Flash->error(__('The organization could not be found. Please, select an existing organization or create a new one.'));
+            return $this->redirect(['action' => 'index']);
+        }
+        
         if ($this->Organizations->delete($organization)) {
             $this->Flash->success(__('The organization has been deleted.'));
         } else {
